viz moved to pyCode
